09_Computer-Hardware-Shop/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Admin==2.0.2
Werkzeug==2.3.7
python-dotenv==1.0.0
